+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Make `opusgain` print previous and new gains when altering files.
 * Make it clearer to Cargo what the licence is.
 * Upgrade to `clap` version 4.
+* Use `wild` for wildcard support on Windows.
 
 ## 0.2.0